[Refactoring] change file names
