+++ SSD/SSD.vcxproj
@@ -130,13 +130,13 @@
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="SSDCommandParser.cpp" />
+    <ClCompile Include="ssd_command_parser.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="SSDCommandParser.h" />
+    <ClInclude Include="ssd_command_parser.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
